# backend/swagger_docs.yaml
openapi: 3.0.0
info:
  title: IDC设备管理系统 API
  version: 1.0.0
  description: |
    IDC设备管理系统后端API文档

    ## 认证说明
    除健康检查接口外，所有接口都需要在请求头中携带JWT Token：
    ```
    Authorization: Bearer <token>
    ```

    ## 通用响应格式
    ### 成功响应
    ```json
    {
      "success": true,
      "data": {...},
      "message": "操作成功"
    }
    ```

    ### 错误响应
    ```json
    {
      "success": false,
      "error": "错误类型",
      "message": "详细描述"
    }
    ```

    ## 分页参数
    列表接口支持 `page` 和 `pageSize` 参数，默认 page=1, pageSize=10
servers:
  - url: http://localhost:8000
    description: 开发环境服务器
security:
  - bearerAuth: []
...
paths:
...
  /api/users:
    get:
      summary: 获取用户列表
      tags: [users]
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: 成功
    post:
      summary: 创建用户
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                roleId:
                  type: string
      responses:
        '201':
          description: 创建成功
  /api/users/{userId}:
    get:
      summary: 获取单个用户
      tags: [users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
    put:
      summary: 更新用户
      tags: [users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 更新成功
    delete:
      summary: 删除用户
      tags: [users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 输入 JWT token